Added admin variable in .env
fix #9

In this fix, I have added admin to .env and also updated the format of
`.env.example`.

In addition, I made little change in `bot.py` to make it more powerfull
in dealing with `.env`.

There is no functionality defined for admin by now. They will come in
the next commit.