Admin can add new products to the bot
fix #14

By uploading a new photo and adding product information as photo
caption, the admin can add new products to the bot.

Each product info is supposed to be uploaded in a separate line and with
key:value format. Look at the following example:

```
----------------
|              |
|    PHOTO     |
|              |
----------------
title: The Product title
size: medium
color: red
price: 50 $
```

It is not necessary to add the detail in any specific order. By now,
title and price will be shown on top and the rest will list after.